// src/models/Person.ts
export interface Person {
  id: number;
  name: string;
  status: string;
  type: string;
  gender: string;
  species: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}